LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/after-scrollable-range-shrinkage-003-expected.txt


FAIL Scroll anchoring properly works after scrollable range shrinkage assert_equals: expected 3008 but got 3626
